# src/GovUK.Dfe.PersonsApi.Client/Generated/swagger.json
{
  "x-generator": "NSwag v14.1.0.0 (NJsonSchema v11.0.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "PersonsApi",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/Constituencies/{constituencyName}/mp": {
      "get": {
        "tags": [
          "Constituencies"
        ],
        "summary": "Retrieve Member of Parliament by constituency name",
        "operationId": "Constituencies_GetMemberOfParliamentByConstituency",
        "parameters": [
          {
            "name": "constituencyName",
            "in": "path",
            "required": true,
            "description": "The constituency name.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "A Person object representing the Member of Parliament.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOfParliament"
                }
              }
            }
          },
          "404": {
            "description": "Constituency not found."
          },
          "400": {
            "description": "Constituency cannot be null or empty."
          }
        }
      }
    },
    "/v1/Constituencies/mps": {
      "post": {
        "tags": [
          "Constituencies"
        ],
        "summary": "Retrieve a collection of Member of Parliament by a collection of constituency names",
        "operationId": "Constituencies_GetMembersOfParliamentByConstituencies",
        "requestBody": {
          "x-name": "request",
          "description": "The request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMembersOfParliamentByConstituenciesQuery"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "A collection of MemberOfParliament objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberOfParliament"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Constituency names cannot be null or empty."
          }
        }
      }
    },
    "/v1/Constituencies/mps/search": {
      "get": {
        "tags": [
          "Constituencies"
        ],
        "summary": "Search for Members of Parliament by either their name or their constituency name",
        "operationId": "Constituencies_SearchMembersOfParliament",
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "description": "A single term matched against both the Member of Parliament name and the constituency name, for example \"John Moore\" or \"London\".",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "A collection of MemberOfParliament objects matching the search term.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberOfParliament"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Search term cannot be null or empty."
          }
        }
      }
    },
    "/v1/Establishments/{urn}/getAssociatedPersons": {
      "get": {
        "tags": [
          "Establishments"
        ],
        "summary": "Retrieve All Members Associated With an Academy by Urn",
        "operationId": "Establishments_GetAllPersonsAssociatedWithAcademyByUrn",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "description": "The URN.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "A Collection of Persons Associated With the Academy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademyGovernance"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Academy not found.\nor\nConstituency not found for the given establishment."
          }
        }
      }
    },
    "/v1/Establishments/{urn}/getMpBySchool": {
      "get": {
        "tags": [
          "Establishments"
        ],
        "summary": "Get Member of Parliament by School (Urn)",
        "operationId": "Establishments_GetMemberOfParliamentBySchoolUrn",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "description": "The URN.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Member of Parliament",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOfParliament"
                }
              }
            }
          },
          "404": {
            "description": "School Not found."
          }
        }
      }
    },
    "/v1/Trusts/{id}/getAssociatedPersons": {
      "get": {
        "tags": [
          "Trusts"
        ],
        "summary": "Retrieve All Members Associated With a Trust by Either UKPRN or TRN",
        "operationId": "Trusts_GetAllPersonsAssociatedWithTrustByTrnOrUkPrn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The identifier (UKPRN or TRN).",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "A Collection of Persons Associated With the Trust.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrustGovernance"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Trust not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MemberOfParliament": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Person"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "constituencyName": {
                "type": "string"
              },
              "constituencyPartyName": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "Person": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string"
          },
          "displayNameWithTitle": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "GetMembersOfParliamentByConstituenciesQuery": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "constituencyNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AcademyGovernance": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Person"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "ukprn": {
                "type": "string",
                "nullable": true
              },
              "urn": {
                "type": "integer",
                "format": "int32",
                "nullable": true
              },
              "dateOfAppointment": {
                "type": "string",
                "nullable": true
              },
              "dateTermOfOfficeEndsEnded": {
                "type": "string",
                "nullable": true
              },
              "appointingBody": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "TrustGovernance": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Person"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "ukprn": {
                "type": "string",
                "nullable": true
              },
              "trn": {
                "type": "string",
                "nullable": true
              },
              "dateOfAppointment": {
                "type": "string",
                "nullable": true
              },
              "dateTermOfOfficeEndsEnded": {
                "type": "string",
                "nullable": true
              },
              "appointingBody": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      }
    }
  }
}